server/weaviate_rag/app/main.py
from fastapi import FastAPI, Request
from pydantic import BaseModel
from app.services.rag import retrieve_answer
from fastapi.middleware.cors import CORSMiddleware

app = FastAPI()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # You can limit this to ["http://localhost:3000"] in production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class ChatRequest(BaseModel):
    question: str

@app.post("/chat")
async def chat_endpoint(req: ChatRequest):
    answer = retrieve_answer(req.question)
    return {"answer": answer}

