server/weaviate_rag/requirements.txt
fastapi
uvicorn
weaviate-client
openai
pdfplumber
tiktoken
python-dotenv
